refactor: separate few components
