+++ src/App.css
@@ -21,17 +21,3 @@
   justify-content: flex-start;
   align-items: flex-start;
 }
-
-.radio-list {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  justify-content: flex-start;
-  align-items: flex-start;
-}
-
-.label {
-  min-width: 200px;
-  max-width: 350px;
-  text-align: left;
-}
