+++ src/App.tsx
@@ -1,21 +1,17 @@
 import { useState } from "react";
 import "./App.css";
 import { FolderTree } from "./components/FolderTree";
-import { pathOptions } from "./constants";
+import { DEFAULT_PATH } from "./constants";
 import FolderTreeService, { StructureNode } from "./services/FolderTreeService";
+import PathesList from "./components/PathesList/PathesList";
+
+const treeService = new FolderTreeService(DEFAULT_PATH);
 
 function App() {
-  const [selectedPath, setSelectedPath] = useState(
-    pathOptions[0].split(", ")[0]
-  );
-  const [treeService] = useState(new FolderTreeService(selectedPath));
-  const [tree, setTree] = useState(treeService.getStructure());
+  const [tree, setTree] = useState(() => treeService.getStructure());
 
   const onChangePath = (path: string) => {
-    const newSelectedPath = path.split(", ")[0];
-
-    setSelectedPath(newSelectedPath);
-    treeService.setStructure(newSelectedPath);
+    treeService.setStructure(path);
     setTree(treeService.getStructure());
   };
 
@@ -33,19 +29,7 @@ function App() {
     <div className="App">
       <h1>Folder Structure</h1>
       <main className="main">
-        <div className="radio-list">
-          {pathOptions.map((path, index) => (
-            <label className="label" key={`${index}-${path}`}>
-              <input
-                type="radio"
-                name="path"
-                checked={selectedPath === path.split(", ")[0]}
-                onChange={() => onChangePath(path)}
-              />
-              {path}
-            </label>
-          ))}
-        </div>
+        <PathesList onChangePath={onChangePath} />
         <FolderTree
           tree={tree}
           addNewItem={addNewItem}
