+++ src/components/PathesList/PathesList.css
@@ -1,0 +1,13 @@
+.radio-list {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  justify-content: flex-start;
+  align-items: flex-start;
+}
+
+.label {
+  min-width: 200px;
+  max-width: 350px;
+  text-align: left;
+}
